kgio can be denied
